fix: working on fixing route guards
