+++ src/app/+state/app.selectors.ts
@@ -1,5 +1,8 @@
 import { createSelector } from '@ngrx/store';
+import { AppStore, ScenariosState } from './app.model';
 import { appFeature } from './app.reducer';
 
-// export const select = createSelector(
-// );
+export const scenariosSelector = createSelector(
+  (state: AppStore) => state.scenarios,
+  (scenarios: ScenariosState) => scenarios
+);
